pyproject.toml
[project]
name = "drawhub-ai-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "openai (>=1.79.0,<2.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
